requirements.txt
Flask==2.3.2
flask-cors==3.0.10
google-generativeai
google-api-python-client
gTTS
